feat(frontend): the webhook layer moves into @agenta/entities, the page into settings-ui
The whole data half moves, not just the view: api, types, atoms, the drawer state
and the test-result handler, all onto @agenta/shared's axios, queryClient and
projectIdAtom. Eleven OSS files import the old service and state paths and keep
working through stubs.

WebhooksPage is antd-free in @agenta/settings-ui; the drawer, the delete dialog
and the one-time secret reveal are slots, since each belongs to the host.
