+++ web/packages/agenta-entities/src/webhook/state.ts
@@ -1,0 +1,9 @@
+import {atom} from "jotai"
+
+import type {WebhookProvider, WebhookSubscription} from "./types"
+
+export const isWebhookDrawerOpenAtom = atom<boolean>(false)
+export const editingWebhookAtom = atom<WebhookSubscription | undefined>(undefined)
+export const createdWebhookSecretAtom = atom<string | null>(null)
+export const selectedProviderAtom = atom<WebhookProvider>("webhook")
+export const webhookToDeleteAtom = atom<WebhookSubscription | null>(null)
