+++ web/packages/agenta-settings-ui/package.json
@@ -16,6 +16,7 @@
     },
     "dependencies": {
         "@agenta/entities": "workspace:../agenta-entities",
+        "@agenta/entity-ui": "workspace:../agenta-entity-ui",
         "@agenta/settings": "workspace:../agenta-settings",
         "@agenta/shared": "workspace:../agenta-shared",
         "@agenta/ui": "workspace:../agenta-ui",
