+++ web/packages/agenta-settings-ui/src/index.ts
@@ -12,3 +12,4 @@ export {AccountPage, type AccountPageProps} from "./AccountPage"
 export {ApiKeysPage, type ApiKeysPageProps} from "./ApiKeysPage"
 export {SecretProviderTable, type SecretProviderTableProps, type ProviderDialogState} from "./secrets/SecretProviderTable"
 export {NamedSecretTable, type NamedSecretTableProps} from "./secrets/NamedSecretTable"
+export {WebhooksPage, type WebhooksPageProps} from "./webhooks/WebhooksPage"
